Deepen command and catalog source ownership
Split the monolithic command parser into domain command modules and move playbook, role, setup, and team profile implementation behind package-owned source modules while preserving root compatibility facades. Add lifecycle invariant tests so prepared handoffs cannot be mistaken for observed execution evidence.

Constraint: User asked to execute all architecture deepening work for the flat src layout in one goal PR.

Rejected: Removing root compatibility modules | existing imports and tests still rely on the stable public facade paths.

Confidence: high

Scope-risk: broad

Directive: Keep root facade modules thin and add new command handlers to the matching domain command module, not commands/main.py.

Tested: PYTHONPATH=tests uv run python -m unittest discover -s tests -v; uv run python -m compileall -q src tests; uv run python -m src.cli docs workflows --check; uv run python -m src.cli harness validate; uv build; git diff --cached --check

Not-tested: Real Hermes runtime activation was not exercised; this remains a local deterministic architecture refactor.
