+++ src/catalogs/__init__.py
@@ -1,0 +1,1 @@
+"""Catalog source-owner modules for OMHM workflow data."""
