+++ src/commands/coding.py
@@ -1,0 +1,257 @@
+from __future__ import annotations
+
+import argparse
+import json
+import sys
+from pathlib import Path
+
+from ..coding_delegation import CODING_EXECUTOR_TARGETS, build_coding_delegation_payload, coding_delegation_record_payload
+from ..ingress import CHAT_SOURCES, extract_message_text, extract_source_metadata
+from ..installer import OmhError
+from ..runtime.artifacts import create_prepared_coding_delegation_run, write_coding_delegation
+from ..wrapper.lifecycle import (
+    CodingLifecycleError,
+    record_codex_dispatch,
+    record_codex_result,
+    record_codex_verification,
+    report_codex_delegation_lifecycle,
+    start_codex_delegation_lifecycle,
+)
+from .common import _chat_input_and_metadata, _explicit_source_metadata, _paths, _print_json, _resolved_executor
+
+
+def cmd_coding_delegate(args: argparse.Namespace) -> int:
+    try:
+        source_metadata: dict[str, str] = {}
+        if args.event_json:
+            raw = (
+                sys.stdin.read()
+                if args.event_json == "-"
+                else Path(args.event_json).expanduser().read_text(encoding="utf-8")
+            )
+            event = json.loads(raw)
+            message = extract_message_text(event)
+            source_metadata = extract_source_metadata(event)
+        elif args.stdin:
+            message = sys.stdin.read().strip()
+        else:
+            message = " ".join(args.message).strip()
+        source_metadata.update(_explicit_source_metadata(args))
+        payload = build_coding_delegation_payload(
+            message,
+            source=args.source,
+            limit=args.limit,
+            include_message=args.include_message,
+            source_metadata=source_metadata,
+            executor_target=_resolved_executor(args, default="generic"),
+        )
+        runtime_skip_reason = _coding_delegate_runtime_skip_reason(payload) if args.record else ""
+        if runtime_skip_reason:
+            payload["runtime"] = {
+                "recorded": False,
+                "reason": runtime_skip_reason,
+                "run_created": False,
+            }
+        elif args.record:
+            delegation = payload["delegation"]
+            if not isinstance(delegation, dict):
+                raise OmhError("coding delegation payload is missing delegation")
+            paths = _paths(args)
+            run = create_prepared_coding_delegation_run(
+                paths,
+                {
+                    "skill": str(delegation["recommended_workflow"]),
+                    "harness": str(delegation["recommended_harness"]),
+                    "trigger": f"coding:{args.source}:{delegation['action']}",
+                    "privacy": "metadata_only",
+                    "inputs_summary": f"{args.source} coding delegation request; message_length={len(message)}",
+                    "outputs_summary": f"prepared {delegation['action']} for {delegation['recommended_workflow']}",
+                    "verification_summary": "prepared_not_observed; executor work is not observed by omh",
+                },
+            )
+            record = write_coding_delegation(
+                paths.runtime_runs_dir / run["run_id"],
+                coding_delegation_record_payload(payload, message, source_metadata=source_metadata),
+            )
+            payload["runtime"] = {"run": run, "coding_delegation": record}
+    except (OSError, json.JSONDecodeError, ValueError) as exc:
+        raise OmhError(str(exc)) from exc
+    _print_json(payload)
+    return 0
+
+
+def _coding_delegate_runtime_skip_reason(payload: dict[str, object]) -> str:
+    selection = payload.get("executor_selection")
+    if isinstance(selection, dict) and selection.get("choice_required") is True:
+        return "executor_choice_required"
+    if payload.get("work_owner_mode") == "prompt_only_handoff":
+        return "prompt_only_handoff_is_wrapper_session_only"
+    if payload.get("work_owner_mode") == "retained_hermes":
+        return "retained_hermes_has_no_executor_handoff"
+    if payload.get("selected_executor_profile") != "codex" or not isinstance(payload.get("executor_handoff"), dict):
+        return "codex_executor_handoff_required_for_runtime_record"
+    return ""
+
+
+def cmd_coding_lifecycle_start(args: argparse.Namespace) -> int:
+    if not args.record:
+        raise OmhError("coding lifecycle start requires --record")
+    if args.executor != "codex":
+        raise OmhError("coding lifecycle is Codex-only in Phase 1; use coding delegate for prompt-only handoffs")
+    try:
+        event_or_message, source_metadata = _chat_input_and_metadata(args)
+        message = extract_message_text(event_or_message)
+        payload = start_codex_delegation_lifecycle(
+            _paths(args),
+            message,
+            source=args.source,
+            source_metadata=source_metadata,
+            limit=args.limit,
+            include_message=args.include_message,
+        )
+    except (OSError, json.JSONDecodeError, ValueError) as exc:
+        raise OmhError(str(exc)) from exc
+    _print_json(payload)
+    return 0
+
+
+def cmd_coding_lifecycle_dispatch(args: argparse.Namespace) -> int:
+    try:
+        _print_json(record_codex_dispatch(_paths(args), args.run_id))
+    except FileNotFoundError as exc:
+        raise OmhError(f"runtime run not found: {args.run_id}") from exc
+    except CodingLifecycleError as exc:
+        raise OmhError(str(exc)) from exc
+    return 0
+
+
+def cmd_coding_lifecycle_result(args: argparse.Namespace) -> int:
+    participants = [item.strip() for item in (args.participants or "").split(",") if item.strip()]
+    try:
+        _print_json(
+            record_codex_result(
+                _paths(args),
+                args.run_id,
+                result=args.result,
+                participants=participants or ["codex"],
+                evidence_refs=args.evidence_ref or [],
+            )
+        )
+    except FileNotFoundError as exc:
+        raise OmhError(f"runtime run not found: {args.run_id}") from exc
+    except CodingLifecycleError as exc:
+        raise OmhError(str(exc)) from exc
+    return 0
+
+
+def cmd_coding_lifecycle_verify(args: argparse.Namespace) -> int:
+    try:
+        _print_json(
+            record_codex_verification(
+                _paths(args),
+                args.run_id,
+                completion_status=args.completion_status,
+                gaps=args.gap or [],
+            )
+        )
+    except FileNotFoundError as exc:
+        raise OmhError(f"runtime run not found: {args.run_id}") from exc
+    except CodingLifecycleError as exc:
+        raise OmhError(str(exc)) from exc
+    return 0
+
+
+def cmd_coding_lifecycle_report(args: argparse.Namespace) -> int:
+    try:
+        _print_json(report_codex_delegation_lifecycle(_paths(args), args.run_id))
+    except FileNotFoundError as exc:
+        raise OmhError(f"runtime run not found: {args.run_id}") from exc
+    return 0
+
+
+def _add_coding_commands(sub) -> None:
+    coding = sub.add_parser("coding")
+    coding_sub = coding.add_subparsers(dest="coding_command", required=True)
+
+    delegate = coding_sub.add_parser("delegate")
+    delegate.add_argument("message", nargs="*", help="Coding task description to prepare for executor delegation.")
+    delegate.add_argument(
+        "--source",
+        choices=CHAT_SOURCES,
+        default="generic",
+        help="Source surface that received the coding request.",
+    )
+    delegate.add_argument("--limit", type=int, default=3, help="Maximum catalog recommendations to include.")
+    delegate.add_argument(
+        "--executor",
+        choices=CODING_EXECUTOR_TARGETS,
+        default=None,
+        help="Optional coding executor target for wrapper handoff payloads.",
+    )
+    delegate.add_argument("--stdin", action="store_true", help="Read the raw coding task from stdin.")
+    delegate.add_argument(
+        "--event-json",
+        default=None,
+        help="Read a Slack/Discord/Hermes-like JSON event from this path, or '-' for stdin.",
+    )
+    delegate.add_argument(
+        "--include-message",
+        action="store_true",
+        help="Include raw message and expanded delegation prompt in stdout for non-logging wrappers.",
+    )
+    delegate.add_argument("--record", action="store_true", help="Record a metadata-only coding delegation artifact under .omh/runtime.")
+    delegate.add_argument("--source-event-id", default="", help="Optional source message/event id to store as metadata.")
+    delegate.add_argument("--channel-ref", default="", help="Optional channel reference to store as metadata.")
+    delegate.add_argument("--user-ref", default="", help="Optional user reference to store as metadata.")
+    delegate.set_defaults(func=cmd_coding_delegate)
+
+    lifecycle = coding_sub.add_parser("lifecycle")
+    lifecycle_sub = lifecycle.add_subparsers(dest="lifecycle_command", required=True)
+
+    lifecycle_start = lifecycle_sub.add_parser("start")
+    lifecycle_start.add_argument("message", nargs="*", help="Coding task description to prepare for Codex lifecycle tracking.")
+    lifecycle_start.add_argument(
+        "--source",
+        choices=CHAT_SOURCES,
+        default="generic",
+        help="Source surface that received the coding request.",
+    )
+    lifecycle_start.add_argument("--limit", type=int, default=3, help="Maximum catalog recommendations to include.")
+    lifecycle_start.add_argument("--executor", choices=CODING_EXECUTOR_TARGETS, default="codex", help="Coding executor target.")
+    lifecycle_start.add_argument("--record", action="store_true", help="Record a metadata-only prepared lifecycle run.")
+    lifecycle_start.add_argument("--stdin", action="store_true", help="Read the raw coding task from stdin.")
+    lifecycle_start.add_argument(
+        "--event-json",
+        default=None,
+        help="Read a Slack/Discord/Hermes-like JSON event from this path, or '-' for stdin.",
+    )
+    lifecycle_start.add_argument(
+        "--include-message",
+        action="store_true",
+        help="Include raw message and expanded executor prompt in stdout for immediate wrapper dispatch.",
+    )
+    lifecycle_start.add_argument("--source-event-id", default="", help="Optional source message/event id to store as metadata.")
+    lifecycle_start.add_argument("--channel-ref", default="", help="Optional channel reference to store as metadata.")
+    lifecycle_start.add_argument("--user-ref", default="", help="Optional user reference to store as metadata.")
+    lifecycle_start.set_defaults(func=cmd_coding_lifecycle_start)
+
+    lifecycle_dispatch = lifecycle_sub.add_parser("dispatch")
+    lifecycle_dispatch.add_argument("--run", dest="run_id", required=True)
+    lifecycle_dispatch.set_defaults(func=cmd_coding_lifecycle_dispatch)
+
+    lifecycle_result = lifecycle_sub.add_parser("result")
+    lifecycle_result.add_argument("--run", dest="run_id", required=True)
+    lifecycle_result.add_argument("--result", choices=("completed", "blocked", "failed"), required=True)
+    lifecycle_result.add_argument("--participants", default="codex")
+    lifecycle_result.add_argument("--evidence-ref", action="append")
+    lifecycle_result.set_defaults(func=cmd_coding_lifecycle_result)
+
+    lifecycle_verify = lifecycle_sub.add_parser("verify")
+    lifecycle_verify.add_argument("--run", dest="run_id", required=True)
+    lifecycle_verify.add_argument("--completion-status", choices=("completed", "blocked", "failed", "unknown"), default="completed")
+    lifecycle_verify.add_argument("--gap", action="append")
+    lifecycle_verify.set_defaults(func=cmd_coding_lifecycle_verify)
+
+    lifecycle_report = lifecycle_sub.add_parser("report")
+    lifecycle_report.add_argument("--run", dest="run_id", required=True)
+    lifecycle_report.set_defaults(func=cmd_coding_lifecycle_report)
