+++ src/profiles/__init__.py
@@ -1,0 +1,1 @@
+"""Profile source-owner modules for OMHM setup and team profiles."""
